+++ R/generate_dummy_mothers_birth_register.R
@@ -12,7 +12,7 @@
 #' @importFrom dplyr mutate arrange desc distinct transmute left_join if_else select rename
 #' @importFrom lubridate as_date dyears make_date days year
 generate_dummy_mothers_birth_register_data <- function(
-    birth_mother_level_data_version="R12v3",
+    birth_mother_level_data_version="R12v1",
     n_patients_minimum = 100,
     seed=13,
     service_sector_data = NULL