+++ data-raw/summary_data.R
@@ -117,26 +117,6 @@ for(folder in summary_data_folders){
 
 
   #
-  # COVARIATES DATA
-  #
-  if(dir.exists(file.path(summary_data_path, folder, "covariates"))){
-    ## covariates scanreport
-
-    col_types = cols(
-      .default = col_character()
-    )
-
-    ScanReport_covaraites <- read_csv(
-      file.path(summary_data_path, folder, "covariates", "ScanReport_covaraites.csv"),
-                                      col_types = col_types)
-
-    ScanReport_covaraites <- ScanReport_covaraites |> mutate_at(vars(starts_with("Frequency...")), as.integer)
-
-    summary_data_versions_list[[folder]]<- list(covariates=list(ScanReport_covaraites = ScanReport_covaraites ))
-
-  }
-
-  #
   # BIRTH MOTHER DATA
   #
   if(dir.exists(file.path(summary_data_path, folder, "birth"))){
@@ -152,7 +132,7 @@ for(folder in summary_data_folders){
 
     ScanReport_birth_mother <- ScanReport_birth_mother |> mutate_at(vars(starts_with("Frequency...")), as.integer)
 
-    summary_data_versions_list[[folder]]<- list(birth_mother=list(ScanReport_birth_mother = ScanReport_birth_mother ))
+    summary_data_versions_list[[folder]][["birth_mother"]]<- list(ScanReport_birth_mother=ScanReport_birth_mother)
 
   }
 
@@ -172,7 +152,7 @@ for(folder in summary_data_folders){
 
     ScanReport_vision <- ScanReport_vision |> mutate_at(vars(starts_with("Frequency...")), as.integer)
 
-    summary_data_versions_list[[folder]]<- list(vision=list(ScanReport_vision = ScanReport_vision ))
+    summary_data_versions_list[[folder]][["vision"]] = list(ScanReport_vision=ScanReport_vision)
 
   }
 
@@ -192,8 +172,7 @@ for(folder in summary_data_folders){
 
     ScanReport_minimum <- ScanReport_minimum |> mutate_at(vars(starts_with("Frequency...")), as.integer)
 
-    summary_data_versions_list[[folder]]<- list(minimum=list(ScanReport_minimum = ScanReport_minimum ))
-
+    summary_data_versions_list[[folder]][["minimum"]] = list(ScanReport_minimum=ScanReport_minimum)
   }